back-end/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.10
pyjwt